// projects/projects/src/api/model/target.ts
/**
 * MaestroHost - SceneService
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Content } from './content';
import { LearningPath } from './learningPath';
import { Scene } from './scene';

export interface Target {
  targetId?: number;
  name?: string;
  description?: string;
  createdAtUTC?: Date;
  scene?: Scene;
  numberOfVisits?: number;
  learningPath?: Array<LearningPath>;
  coordinateX?: number;
  coordinateY?: number;
  contents?: Array<Content>;
  readonly sceneId?: number;
}